ปิดการใช้งาน OsmDroid ใน Class ที่เกี่ยวข้องกับแผนที่ที่ใช้ OpenStreetMap

+++ app/src/main/java/th/in/ffc/map/overlay/FGOverlay.java
@@ -44,8 +44,8 @@ public class FGOverlay extends Overlay {
     @Override
     public boolean onLongPress(MotionEvent e, MapView mapView) {
 
-        IGeoPoint geoPoint = mapView.getProjection().fromPixels(e.getX(),
-                e.getY());
+        IGeoPoint geoPoint = mapView.getProjection().fromPixels((int) e.getX(),
+                (int) e.getY());
 
         if (this.fgSystemManager.getFGDatabaseManager().getAvailable().size() == 0) {
             AlertDialog alertDialogAllHouseHasBeenCreated = this.fgSystemManager
